# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Omnichannel gateway, campaign engine, Direct Universal SIP Trunking PBX, MCP (Model Context Protocol), Unified RAG Router, Recall.ai Meeting Bots \u0026 Evolution Go Proxy",
        "title": "Patter Engine Gateway \u0026 Campaign Engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/lynxflow/patter-go"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/gateway/whatsapp/connect": {
            "post": {
                "description": "Inscrit une nouvelle session WhatsApp sous la marque blanche du Tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway - WhatsApp"
                ],
                "summary": "Connecter une session WhatsApp (Evolution Go)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du Tenant White-Label",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Information de la session",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ConnectWhatsAppRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.QRCodeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/whatsapp/contact/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway - WhatsApp"
                ],
                "summary": "Envoyer une carte de contact VCard WhatsApp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du Tenant White-Label",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Paramètres contact",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendContactRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SendMessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/whatsapp/location/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway - WhatsApp"
                ],
                "summary": "Envoyer une localisation GPS WhatsApp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du Tenant White-Label",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Paramètres localisation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SendMessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/whatsapp/media/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway - WhatsApp"
                ],
                "summary": "Envoyer un média WhatsApp (Image, Vidéo, Document, Audio, Sticker)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du Tenant White-Label",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Paramètres média",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendMediaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SendMessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/whatsapp/message/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway - WhatsApp"
                ],
                "summary": "Envoyer un message WhatsApp texte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du Tenant White-Label",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Contenu du message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SendMessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/whatsapp/qrcode": {
            "get": {
                "description": "Renvoie le QR Code Base64 pour authentification marque blanche",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway - WhatsApp"
                ],
                "summary": "Obtenir le QR Code d'une session WhatsApp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du Tenant White-Label",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nom de la session",
                        "name": "session",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.QRCodeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/whatsapp/sessions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway - WhatsApp"
                ],
                "summary": "Lister les sessions WhatsApp du Tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du Tenant White-Label",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.SessionStatusResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "core.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.ConnectWhatsAppRequest": {
            "type": "object",
            "required": [
                "session_name"
            ],
            "properties": {
                "phone_number": {
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.QRCodeResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "pairing_code": {
                    "type": "string"
                },
                "qrcode_base64": {
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                },
                "status": {
                    "description": "\"SCAN_QR_CODE\", \"WORKING\", \"CONNECTING\"",
                    "type": "string"
                }
            }
        },
        "dto.SendContactRequest": {
            "type": "object",
            "required": [
                "contact_name",
                "phone",
                "recipient",
                "session_name"
            ],
            "properties": {
                "contact_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                }
            }
        },
        "dto.SendLocationRequest": {
            "type": "object",
            "required": [
                "latitude",
                "longitude",
                "recipient",
                "session_name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                }
            }
        },
        "dto.SendMediaRequest": {
            "type": "object",
            "required": [
                "media_type",
                "media_url",
                "recipient",
                "session_name"
            ],
            "properties": {
                "caption": {
                    "type": "string"
                },
                "media_type": {
                    "description": "\"image\", \"video\", \"document\", \"audio\", \"sticker\"",
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                }
            }
        },
        "dto.SendMessageRequest": {
            "type": "object",
            "required": [
                "message",
                "recipient",
                "session_name"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "recipient": {
                    "description": "E.164 format",
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                }
            }
        },
        "dto.SendMessageResponse": {
            "type": "object",
            "properties": {
                "message_id": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                },
                "status": {
                    "description": "\"SENT\", \"PENDING\", \"FAILED\"",
                    "type": "string"
                }
            }
        },
        "dto.SessionStatusResponse": {
            "type": "object",
            "properties": {
                "connected": {
                    "type": "boolean"
                },
                "phone_number": {
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                },
                "state": {
                    "description": "\"WORKING\", \"DISCONNECTED\", \"CONNECTING\", \"PAIRING\"",
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
...
    }
}